for some last few comments, i have been struggling with the unittesting only to find out that i blanked out on how to create them. I watched a quick video and then i remembered how to do it. so yay. today, i finished up unittesting and fixed some errors within the testing that wouldn't allow the program to run properly.
i made some comments on my code where on certain parts i struggled and went online for some help. I'm still a bit shaky on javascript and c, but I am much more confident with html now.

i believe i have 4 unittests in  folder where it tests 4 things and they did pass the tests. during the writing process of the tests, i went to chatgpt to ask why there were errors, but it was how i wrote the message and it didn't line up properly. once i fixed the message, it was working and i was sitting there for a good 30 minutes wondering why it wasn't working, but that happened.